+++ node-api/app.js
@@ -3,13 +3,15 @@ const app = express();
 const bodyParser = require('body-parser');
 
 // import the route
-const routes = require('./routes/routes');
+const publicRoutes = require('./routes/routes');
+const authRoutes = require('./admin/adminRoutes/routes');
 
 // body parser config
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: true }));
 
 // Registering the routes
-app.use('/', routes);
+app.use('/', publicRoutes);
+app.use('/admin', authRoutes)
 
 module.exports = app;
