+++ node-api/package.json
@@ -12,8 +12,10 @@
   "dependencies": {
     "@azure/data-tables": "^13.2.2",
     "@azure/storage-blob": "^12.16.0",
+    "bcrypt": "^5.1.1",
     "dotenv": "^16.3.1",
     "express": "^4.18.2",
+    "jsonwebtoken": "^9.0.2",
     "multer": "^1.4.5-lts.1",
     "uuid": "^9.0.1"
   },
